Moved test files in better location
Apparently, some test files where created during the last revert.
